requirements.txt
pdfplumber==0.10.3
reportlab==4.0.8
Pillow==10.2.0
colorama==0.4.6
tqdm==4.66.1
python-dateutil==2.8.2
tabulate==0.9.0
openpyxl==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
